backend/requirements.txt
# FastAPI 和相关依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库相关
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 配置管理
pydantic==2.5.0
pydantic-settings==2.0.3
email-validator==2.1.0

# HTTP 客户端
httpx==0.25.2
aiohttp==3.9.1

# 工具库
python-dateutil==2.8.2
pytz==2023.3
openpyxl==3.1.2
Pillow==10.1.0

# 日志和监控
loguru==0.7.2
prometheus-client==0.19.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# 邮件发送
emails==0.6
jinja2==3.1.2

# 任务队列 (可选)
celery==5.3.4
flower==2.0.1

# 文档生成
mkdocs==1.5.3
mkdocs-material==9.4.8